Fix circular import by removing import of DingoDataset from dingo.core.utils.
