+++ dingo/core/utils/__init__.py
@@ -1,5 +1,4 @@
 from .torchutils import *
 from .trainutils import *
-from .dataset_utils import *
 from .gnpeutils import *
-from .misc_utils import get_version
+from .misc import *
