--- dingo/core/utils/misc_utils.py
@@ -1,6 +1,0 @@
-def get_version():
-    try:
-        from dingo import __version__
-        return __version__
-    except ImportError:
-        return None
